nextWriteValue will be set to NextValue --> value

+++ io.openems.edge.consolinno.leaflet.maindevice/src/io/openems/edge/consolinno/leaflet/maindevice/doubleuart/Config.java
@@ -25,7 +25,7 @@ import org.osgi.service.metatype.annotations.Option;
     String informationType() default "OnOff";
 
     @AttributeDefinition(name = "GPIO ", description = "What GPIO u want to call (0-7) is possible.")
-    byte pinPosition() default 0;
+    int pinPosition() default 0;
 
     @AttributeDefinition(name = "Spi Channel", description = "Channel Address of DoubleUART")
     int spiChannel() default 8;
